Move the logic to a task but run in through the processor cause the tests expect that right now, moving the tests over next
